added a new state exchange_card for ais, in order for them not to exchange twice their armies
this was happening for ais owning a continent, who received twice (or
more the event place_armies, and thus tried to exchange cards more than
once. and since controller does not (yet) control that player has the
cards s/he wants to exchange... this resulted in free bonus armies for
the ai!